Several options were logically ints instead of doubles. So, I changed them and fixed the methods that used them.

+++ ImageTracerNet/OptionTypes/Blur.cs
@@ -5,7 +5,7 @@ namespace ImageTracerNet.OptionTypes
     [Serializable]
     public class Blur
     {
-        public double BlurRadius { get; set; } = 0f;
+        public int BlurRadius { get; set; } = 0;
         public double BlurDelta { get; set; } = 20f;
     }
 }
